More supply updates (#546)
* More supply updates

* Tiny ui fix

* Fix logs

* Fix cooldown display

* Changes favor requirement on some traders

* More tweaks & fixes

* Platinum for sale

* Reverse-list instead of reverse-range

* Fix logs display again

* Wizard item worths

* Display balance in the console

* Payrolls test

* Hidden trading stations tiny tweak

* Faction link cooldown

* Grown plants and gas values

* Market crash effect!

* Fix issue with SMG mags

* Changes how market crash works

* Fix gas proc

* Fix grown plants value

* Fix atmos worth again

* Atmos tweaks

* Canisters don't late init their gases

* Fix attempt

* Fix please

* Pump fix

* Supermatter is now sold in hidden inv of engi trader

* Payrolls update

* Artifact values

* Fix

* More trading stations

* Fix duplicate access name

* Fix

* Some tweaks

Organ value depends on species rarity

* Fix payrolls

* Engraver uses considerably more power now

* Dominion trader price change

* Tweak

* Slime extract values

* Adds more exempt types

* External organs cost scales

* Seeds price fix

* Auto-login at 0 security level

* Dressup can be used on clientless mobs and ghosts

* Organ price tweaks

* Trading beacon circuits & designs

* Construct states for trade beacons

* Fix spellbook value

* Rockets have higher weight

+++ code/controllers/subsystems/supply.dm
@@ -3,6 +3,8 @@ SUBSYSTEM_DEF(supply)
 	name = "Supply"
 	priority = SS_PRIORITY_SUPPLY
 	flags = SS_NO_FIRE
+	runlevels = RUNLEVEL_GAME|RUNLEVEL_POSTGAME
+	wait = 30 MINUTES
 
 	/// Budget for trading stations with spawn_always at FALSE
 	var/trade_stations_budget = 5
@@ -10,8 +12,6 @@ SUBSYSTEM_DEF(supply)
 	var/list/all_trading_stations = list()
 	/// All visible and available stations
 	var/list/visible_trading_stations = list()
-	/// All hidden/unavailable stations
-	var/list/hidden_trading_stations = list()
 	/// All factions in the world; Assoc list Name = Datum.
 	/// Trading stations get the datums from here
 	var/list/factions = list()
@@ -34,6 +34,23 @@ SUBSYSTEM_DEF(supply)
 	var/handling_fee = 0.1
 	var/order_queue_id = 0
 	var/list/order_queue = list()
+
+	// For exporting
+	/// Maximum price of an atom before it becomes affected by an export modifier
+	var/export_maximum = 25000
+	var/list/export_modifier = list()
+	var/list/export_counter = list()
+	/// Paths of atoms that are entirely unaffected by export price modifiers
+	var/list/export_modifier_exempt_types = list(
+		/obj/item/ore,
+		/obj/item/stack,
+		/obj/item/reagent_containers,
+		/obj/item/slime_extract,
+		/obj/item/ammo_casing,
+		/obj/machinery/artifact,
+		/obj/machinery/portable_atmospherics/canister,
+		/mob/living/carbon/slime,
+		)
 
 /datum/controller/subsystem/supply/Initialize()
 	. = ..()
@@ -67,6 +84,14 @@ SUBSYSTEM_DEF(supply)
 	DeInitTradeStations()
 	. = ..()
 
+/datum/controller/subsystem/supply/fire()
+	for(var/datum/money_account/A in all_money_accounts)
+		A.PayrollTick()
+	for(var/A in export_counter)
+		var/reduction = max(export_maximum, export_counter[A] * 0.5)
+		export_counter[A] = max(0, export_counter[A] -= reduction)
+	ExportCounterCheck()
+
 /datum/controller/subsystem/supply/proc/GetFaction(fac)
 	if(!(fac in factions))
 		return null
@@ -96,7 +121,6 @@ SUBSYSTEM_DEF(supply)
 	for(var/datum/trading_station/TS in all_trading_stations)
 		TS.RegainTradeStationsBudget()
 		qdel(TS)
-		hidden_trading_stations -= TS
 		visible_trading_stations -= TS
 		all_trading_stations -= TS
 
@@ -275,7 +299,7 @@ SUBSYSTEM_DEF(supply)
 
 	return order_queue_slot
 
-/datum/controller/subsystem/supply/proc/PurchaseOrder(obj/machinery/trade_beacon/receiving/beacon, order_id)
+/datum/controller/subsystem/supply/proc/PurchaseOrder(obj/machinery/trade_beacon/receiving/beacon, order_id, used_faction = FACTION_INDEPENDENT)
 	if(QDELETED(beacon) || !beacon || !order_id)
 		return
 
@@ -289,12 +313,12 @@ SUBSYSTEM_DEF(supply)
 		var/total_cost = order["cost"] + order["fee"]
 		var/is_requestor_master = (requesting_account == master_account) ? TRUE : FALSE
 
-		Buy(beacon, master_account, shopping_list, !is_requestor_master, requesting_account.owner_name)
+		Buy(beacon, master_account, shopping_list, !is_requestor_master, requesting_account.owner_name, used_faction)
 		if(!is_requestor_master)
 			requesting_account.transfer(master_account, total_cost, "Order Request")
-		CreateLogEntry("Order", requesting_account.owner_name, viewable_contents, total_cost)
+		CreateLogEntry("Order", requesting_account.owner_name, used_faction, viewable_contents, total_cost)
 
-/datum/controller/subsystem/supply/proc/Buy(obj/machinery/trade_beacon/receiving/senderBeacon, datum/money_account/account, list/shopList, is_order = FALSE, buyer_name = null)
+/datum/controller/subsystem/supply/proc/Buy(obj/machinery/trade_beacon/receiving/senderBeacon, datum/money_account/account, list/shopList, is_order = FALSE, buyer_name = null, used_faction = FACTION_INDEPENDENT)
 	if(QDELETED(senderBeacon) || !istype(senderBeacon) || !account || !RecursiveLen(shopList))
 		return FALSE
 
@@ -348,12 +372,15 @@ SUBSYSTEM_DEF(supply)
 	if(count_of_all > 1)
 		invoice_location = C
 
-	CreateLogEntry("Shipping", account.owner_name, order_contents_info, price_for_all, FALSE, invoice_location)
+	CreateLogEntry("Shipping", account.owner_name, used_faction, order_contents_info, price_for_all, FALSE, invoice_location)
 	account.withdraw(price_for_all, "Purchase", "Trade Network")
 	return TRUE
 
-/datum/controller/subsystem/supply/proc/Export(obj/machinery/trade_beacon/sending/senderBeacon, datum/money_account/moneyAccount)
+/datum/controller/subsystem/supply/proc/Export(obj/machinery/trade_beacon/sending/senderBeacon, datum/money_account/moneyAccount, used_faction = FACTION_INDEPENDENT)
 	if(QDELETED(senderBeacon))
+		return
+
+	if(!senderBeacon.StartExport())
 		return
 
 	var/invoice_contents_info
@@ -373,12 +400,20 @@ SUBSYSTEM_DEF(supply)
 
 		// We go backwards, so it'll be innermost objects sold first
 		for(var/atom/movable/item in reverseRange(contents_incl_self))
-			var/item_price = get_value(item)
+			if(istype(item, /obj/screen))
+				continue
+
+			var/item_price = GetExportValue(item)
 			var/export_value = item_price
 
 			if(export_value)
 				invoice_contents_info += "<li>[item.name]</li>"
 				cost += export_value
+				if(!(is_path_in_list(item.type, export_modifier_exempt_types)))
+					if(!(item.type in export_counter))
+						export_counter[item.type] = 0
+					if(export_counter[item.type] < export_maximum * 2)
+						export_counter[item.type] += round(min(export_value, export_maximum * 0.25))
 				//SEND_SIGNAL(src, COMSIG_TRADE_BEACON, item)
 				qdel(item)
 				++export_count
@@ -392,31 +427,44 @@ SUBSYSTEM_DEF(supply)
 		if(export_count > 100)
 			break
 
-	senderBeacon.StartExport()
+	ExportCounterCheck()
+
 	moneyAccount.deposit(cost, "Export", "Trade Network")
 
 	if(invoice_contents_info)	// If no info, then nothing was exported
-		CreateLogEntry("Export", moneyAccount.owner_name, invoice_contents_info, cost, FALSE, get_turf(senderBeacon))
+		CreateLogEntry("Export", moneyAccount.owner_name, used_faction, invoice_contents_info, cost, FALSE, get_turf(senderBeacon))
 	return TRUE
+
+/datum/controller/subsystem/supply/proc/GetExportValue(atom/A)
+	. = get_value(A)
+	if(A.type in export_modifier)
+		. *= export_modifier[A.type]
+
+/datum/controller/subsystem/supply/proc/ExportCounterCheck()
+	for(var/A in export_counter)
+		if(export_counter[A] > export_maximum)
+			export_modifier[A] = clamp(round(export_maximum / export_counter[A], 0.01), 0.5, 2.0)
+		else if(A in export_modifier)
+			export_modifier -= A
 
 // Logging
 
-/datum/controller/subsystem/supply/proc/CreateLogEntry(type, ordering_account, contents, total_paid, create_invoice = FALSE, invoice_location = null)
+/datum/controller/subsystem/supply/proc/CreateLogEntry(type, ordering_account, assoc_faction = FACTION_INDEPENDENT, contents, total_paid, create_invoice = FALSE, invoice_location = null)
 	var/log_id
 
 	switch(type)
 		if("Shipping")
 			log_id = "[++shipping_invoice_number]-S"
-			shipping_log.Add(list(list("id" = log_id, "ordering_acct" = ordering_account, "contents" = contents, "total_paid" = total_paid, "time" = time2text(world.time, "hh:mm"))))
+			shipping_log.Add(list(list("id" = log_id, "ordering_acct" = ordering_account, "assoc_faction" = assoc_faction, "contents" = contents, "total_paid" = total_paid, "time" = time2text(world.time, "hh:mm"))))
 		if("Export")
 			log_id = "[++export_invoice_number]-E"
-			export_log.Add(list(list("id" = log_id, "ordering_acct" = ordering_account, "contents" = contents, "total_paid" = total_paid, "time" = time2text(world.time, "hh:mm"))))
+			export_log.Add(list(list("id" = log_id, "ordering_acct" = ordering_account, "assoc_faction" = assoc_faction, "contents" = contents, "total_paid" = total_paid, "time" = time2text(world.time, "hh:mm"))))
 		if("Special Offer")
 			log_id = "[++offer_invoice_number]-SO"
-			offer_log.Add(list(list("id" = log_id, "ordering_acct" = ordering_account, "contents" = contents, "total_paid" = total_paid, "time" = time2text(world.time, "hh:mm"))))
+			offer_log.Add(list(list("id" = log_id, "ordering_acct" = ordering_account, "assoc_faction" = assoc_faction, "contents" = contents, "total_paid" = total_paid, "time" = time2text(world.time, "hh:mm"))))
 		if("Order")
 			log_id = "[++order_number]-O"
-			order_log.Add(list(list("id" = log_id, "ordering_acct" = ordering_account, "contents" = contents, "total_paid" = total_paid, "time" = time2text(world.time, "hh:mm"))))
+			order_log.Add(list(list("id" = log_id, "ordering_acct" = ordering_account, "assoc_faction" = assoc_faction, "contents" = contents, "total_paid" = total_paid, "time" = time2text(world.time, "hh:mm"))))
 		else
 			return
